su24-cs411-team002-dropTable-main/Project/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloud-sql-python-connector==1.11.0
cryptography==43.0.0
Flask==3.0.3
frozenlist==1.4.1
google-auth==2.32.0
gunicorn==22.0.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
multidict==6.0.5
packaging==24.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
PyMySQL==1.1.1
requests==2.32.3
rsa==4.9
urllib3==2.2.2
Werkzeug==3.0.3
yarl==1.9.4
